feat: move recovery mechanism in separated folder
