+++ src/pages/ValidatorOnboarding/KeyGeneration/RecoveryMechanism/RecoveryMechanism.stories.ts
@@ -1,24 +1,23 @@
 import type { Meta, StoryObj } from '@storybook/react'
 
-import RecoveryMechanismCard from './RecoveryMechanismCard'
-import { KEYSTORE_FILES } from '../../../constants'
+import RecoveryMechanism from './RecoveryMechanism'
+import { KEYSTORE_FILES } from '../../../../constants'
 
 const meta = {
-  title: 'ValidatorOnboarding/RecoveryMechanismCard',
-  component: RecoveryMechanismCard,
+  title: 'ValidatorOnboarding/RecoveryMechanism',
+  component: RecoveryMechanism,
   parameters: {
     layout: 'centered',
   },
   tags: ['autodocs'],
   argTypes: {},
-} satisfies Meta<typeof RecoveryMechanismCard>
+} satisfies Meta<typeof RecoveryMechanism>
 
 export default meta
 type Story = StoryObj<typeof meta>
 
 export const Page: Story = {
   args: {
-    value: KEYSTORE_FILES,
     recoveryMechanism: KEYSTORE_FILES,
     handleRecMechanismChange: () => {},
   },
