+++ src/pages/ValidatorOnboarding/KeyGeneration/RecoveryMechanism/RecoveryMechanism.tsx
@@ -2,7 +2,7 @@ import { Text } from '@status-im/components'
 import { XStack, YStack } from 'tamagui'
 
 import RecoveryMechanismCard from './RecoveryMechanismCard'
-import { BOTH_KEY_AND_RECOVERY, KEYSTORE_FILES, RECOVERY_PHRASE } from '../../../constants'
+import { BOTH_KEY_AND_RECOVERY, KEYSTORE_FILES, RECOVERY_PHRASE } from '../../../../constants'
 
 type RecoveryMechanismProps = {
   recoveryMechanism: string
