+++ src/pages/ValidatorOnboarding/KeyGeneration/RecoveryMechanism/RecoveryMechanismCard.stories.ts
@@ -1,23 +1,24 @@
 import type { Meta, StoryObj } from '@storybook/react'
 
-import RecoveryMechanism from './RecoveryMechanism'
-import { KEYSTORE_FILES } from '../../../constants'
+import RecoveryMechanismCard from './RecoveryMechanismCard'
+import { KEYSTORE_FILES } from '../../../../constants'
 
 const meta = {
-  title: 'ValidatorOnboarding/RecoveryMechanism',
-  component: RecoveryMechanism,
+  title: 'ValidatorOnboarding/RecoveryMechanismCard',
+  component: RecoveryMechanismCard,
   parameters: {
     layout: 'centered',
   },
   tags: ['autodocs'],
   argTypes: {},
-} satisfies Meta<typeof RecoveryMechanism>
+} satisfies Meta<typeof RecoveryMechanismCard>
 
 export default meta
 type Story = StoryObj<typeof meta>
 
 export const Page: Story = {
   args: {
+    value: KEYSTORE_FILES,
     recoveryMechanism: KEYSTORE_FILES,
     handleRecMechanismChange: () => {},
   },
